# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.PracticeRequestDto:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  dto.PracticeResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  membership_plan.PlanRequestDto:
    properties:
      amt_periods:
        type: integer
      membership_id:
        type: string
      name:
        type: string
      payment_frequency:
        type: string
      price:
        type: integer
    required:
    - membership_id
    - price
    type: object
  membership_plan.PlanResponse:
    properties:
      amt_periods:
        type: integer
      id:
        type: string
      membership_id:
        type: string
      name:
        type: string
      payment_frequency:
        type: string
      price:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
paths:
...
  /membership-plan:
    get:
      consumes:
      - application/json
      description: Get membership plans by membership HubSpotId
      parameters:
      - description: Filter by customer ID
        in: query
        name: customerId
        type: string
      - description: Filter by membership ID
        in: query
        name: membershipId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetMemberships of membership plans retrieved successfully
          schema:
            items:
              $ref: '#/definitions/membership_plan.PlanResponse'
            type: array
        "400":
          description: 'Bad Request: Invalid membership HubSpotId'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Membership plans not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get membership plans by membership HubSpotId
      tags:
      - membership-plans
    post:
      consumes:
      - application/json
      description: Create a new membership plan
      parameters:
      - description: Membership plan details
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/membership_plan.PlanRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Membership plan created successfully
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new membership plan
      tags:
      - membership-plans
  /membership-plan/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a membership plan by HubSpotId
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: Membership plan deleted successfully'
        "400":
          description: 'Bad Request: Invalid HubSpotId'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Membership plan not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a membership plan
      tags:
      - membership-plans
    put:
      consumes:
      - application/json
      description: Update a membership plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Membership plan details
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/membership_plan.PlanRequestDto'
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: Membership plan updated successfully'
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Membership plan not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update a membership plan
      tags:
      - membership-plans
...
  /practices:
    get:
      consumes:
      - application/json
      description: Get a list of practices
      parameters:
      - description: Filter by practice name
        in: query
        name: name
        type: string
      - description: Filter by practice description
        in: query
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetMemberships of practices retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.PracticeResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a list of practices
      tags:
      - practices
    post:
      consumes:
      - application/json
      description: Create a new practice
      parameters:
      - description: Practice details
        in: body
        name: practice
        required: true
        schema:
          $ref: '#/definitions/dto.PracticeRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Practice created successfully
          schema:
            $ref: '#/definitions/dto.PracticeResponse'
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new practice
      tags:
      - practices
  /practices/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a practice by HubSpotId
      parameters:
      - description: Practice HubSpotId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: Practice deleted successfully'
        "400":
          description: 'Bad Request: Invalid HubSpotId'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Practice not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a practice
      tags:
      - practices
    put:
      consumes:
      - application/json
      description: Update a practice
      parameters:
      - description: Practice HubSpotId
        in: path
        name: id
        required: true
        type: string
      - description: Practice details
        in: body
        name: practice
        required: true
        schema:
          $ref: '#/definitions/dto.PracticeRequestDto'
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: Practice updated successfully'
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Practice not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update a practice
      tags:
      - practices
  /practices/{name}:
    get:
      consumes:
      - application/json
      description: Get a practice by name
      parameters:
      - description: Practice Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Practice retrieved successfully
          schema:
            $ref: '#/definitions/dto.PracticeResponse'
        "400":
          description: 'Bad Request: Invalid Name'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Practice not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a practice by name
      tags:
      - practices
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"